renaming the compiler folders
