+++ lib/browser/compiler/index.js
@@ -1,5 +1,5 @@
 import './wrap/prefix'
 import './parsers'
-import '../../helpers/compiler-shared'
+import '../../shared/compiler'
 import './core'
 import './wrap/suffix'
