+++ lib/server/cli.js
