+++ lib/server/compiler.js
@@ -1,4 +1,4 @@
-var compilerPath = require('path').join(__dirname, '../helpers/compiler-shared.js')
+var compilerPath = require('path').join(__dirname, '../shared/compiler.js')
 
 this.riot = require(process.env.RIOT || '../../riot')
 this.riot.parsers = require('./parsers')
