+++ lib/server/sdom.js
