+++ lib/shared/compiler.js
